Moved ClientSession
