+++ BranchSDK/src/BranchIO/Util/Storage.h
@@ -1,7 +1,7 @@
 // Copyright (c) 2019 Branch Metrics, Inc.
 
-#ifndef BRANCHIO_STORAGE_H__
-#define BRANCHIO_STORAGE_H__
+#ifndef BRANCHIO_UTIL_STORAGE_H__
+#define BRANCHIO_UTIL_STORAGE_H__
 
 #ifdef WIN32
 #include "BranchIO/WindowsStorage.h"
@@ -17,4 +17,4 @@ typedef UnixStorage Storage;
 #endif  // WIN32
 }  // namespace BranchIO
 
-#endif  // BRANCHIO_STORAGE_H__
+#endif  // BRANCHIO_UTIL_STORAGE_H__
