+++ BranchSDK/src/BranchIO/Util/WindowsStorage.h
@@ -1,7 +1,7 @@
 // Copyright (c) 2019 Branch Metrics, Inc.
 
-#ifndef BRANCHIO_WINDOWSSTORAGE_H_
-#define BRANCHIO_WINDOWSSTORAGE_H_
+#ifndef BRANCHIO_UTIL_WINDOWSSTORAGE_H_
+#define BRANCHIO_UTIL_WINDOWSSTORAGE_H_
 
 #include <Poco/Mutex.h>
 
@@ -92,4 +92,4 @@ class WindowsStorage : public virtual IStorage {
 
 }  // namespace BranchIO
 
-#endif  // BRANCHIO_WINDOWSSTORAGE_H_
+#endif  // BRANCHIO_UTIL_WINDOWSSTORAGE_H_
